extract traceUtil
